+++ src/modbus_client.rs
@@ -29,14 +29,14 @@ use std::io::Error;
 ///
 pub trait ModbusClient {
     fn read_bits(&self, address: u16, num: u16, dest: &mut [u8]) -> Result<i32>;
-    fn read_input_bits(&self, address: i32, num: i32, dest: &mut [u8]) -> Result<i32>;
+    fn read_input_bits(&self, address: u16, num: u16, dest: &mut [u8]) -> Result<i32>;
     fn read_registers(&self, address: i32, num: i32, dest: &mut [u16]) -> Result<i32>;
     fn read_input_registers(&self, address: i32, num: i32, dest: &mut [u16]) -> Result<i32>;
     fn report_slave_id(&self, max_dest: usize, dest: &mut [u8]) -> Result<i32>;
     fn write_bit(&self, address: u16, status: bool) -> Result<()>;
     fn write_bits(&self, address: u16, num: u16, src: &[u8]) -> Result<i32>;
     fn write_register(&self, address: i32, value: i32) -> Result<()>;
-    fn write_registers(&self, address: i32, num: i32, src: &[u16]) -> Result<i32>;
+    fn write_registers(&self, address: u16, num: u16, src: &[u16]) -> Result<i32>;
     fn write_and_read_registers(&self, write_address: i32, write_num: i32, src: &[u16], read_address: i32,
                                 read_num: i32, dest: &mut [u16])
                                 -> Result<i32>;
@@ -113,13 +113,13 @@ impl ModbusClient for Modbus {
     ///
     /// assert!(modbus.read_input_bits(0, 1, &mut dest).is_ok());
     /// ```
-    fn read_input_bits(&self, address: i32, num: i32, dest: &mut [u8]) -> Result<i32> {
-        if num > dest.len() as i32 {
+    fn read_input_bits(&self, address: u16, num: u16, dest: &mut [u8]) -> Result<i32> {
+        if num > dest.len() as u16 {
             bail!(ErrorKind::TooManyData("Too many discrete inputs requested"));
         }
 
         unsafe {
-            match libmodbus_sys::modbus_read_input_bits(self.ctx, address as c_int, num, dest.as_mut_ptr()) {
+            match libmodbus_sys::modbus_read_input_bits(self.ctx, address as c_int, num as c_int, dest.as_mut_ptr()) {
                 -1 => bail!(Error::last_os_error()),
                 len => Ok(len),
             }
@@ -387,9 +387,9 @@ impl ModbusClient for Modbus {
     ///
     /// assert_eq!(modbus.write_registers(address, 1, &tab_bytes).unwrap(), 1);
     /// ```
-    fn write_registers(&self, address: i32, num: i32, src: &[u16]) -> Result<i32> {
+    fn write_registers(&self, address: u16, num: u16, src: &[u16]) -> Result<i32> {
         unsafe {
-            match libmodbus_sys::modbus_write_registers(self.ctx, address as c_int, num, src.as_ptr()) {
+            match libmodbus_sys::modbus_write_registers(self.ctx, address as c_int, num as c_int, src.as_ptr()) {
                 -1 => bail!(Error::last_os_error()),
                 num => Ok(num),
             }
